Move clockinfo config to settings
