+++ apps/mtnclock/app.js
@@ -5,8 +5,6 @@ if (! Array.isArray(data.rows)) {
   require("Storage").writeJSON("mtnclock.json", data);
 }
 
-let showingSettings = false;
-let clockinfosSettings = [];
 let clockinfosMain = {};
 let clockinfos = require("clock_info").load();
 
@@ -236,7 +234,6 @@ function draw(color) {
 }
 
 function setWeather() {
-  if (showingSettings) return;
   var a = {};
   //clear day/night is default weather
   if ((data.code >= 800 && data.code <=802) || data.code == undefined) {
@@ -398,60 +395,6 @@ global.GB = (event) => {
   if (_GB) setTimeout(_GB, 0, event);
 };
 
-function drawClkinfoSettings() {
-  if (drawTimeout) clearTimeout(drawTimeout);
-  g.clear();
-  g.setColor(g.theme.fg);
-  g.setFont("Vector", py(10)).setFontAlign(-1, -1).drawString("<Back", px(2), py(9));
-  g.drawRect(1, 1, px(33), px(25)-1);
-  g.setFont("Vector", py(10)).setFontAlign(0, -1).drawString("-", px(50), py(9));
-  g.drawRect(px(33), 1, px(67), px(25)-1);
-  g.setFont("Vector", py(10)).setFontAlign(0, -1).drawString("+", px(83), py(9));
-  g.drawRect(px(67), 1, px(100)-1, px(25)-1);
-
-  data.rows.forEach(function(row, r) {
-    let a = row.menuA;
-    let b = row.menuB;
-    let ci = clockinfos[a].items[b];
-    if (clockinfosMain[a] && clockinfosMain[a][b]) {
-      clockinfosMain[a][b] = false;
-      ci.hide();
-      ci.removeListener("redraw", clockinfoRedraw);
-    }
-    addClockinfo(r)
-  });
-  clockinfosMain = {};
-}
-
-function addClockinfo(r) {
-  let dr = data.rows[r];
-	// Check if the saved clockinfo indices still exist
-	let ma = (dr && dr.menuA && clockinfos[dr.menuA]) ? dr.menuA : 0;
-	let mb = (ma && dr && dr.menuB && clockinfos[ma].items[dr.menuB]) ? dr.menuB : 0;
-  clockinfosSettings[r] = require("clock_info").addInteractive(clockinfos, {
-    x : 2, y: py((r+1)*25)+1, w: px(100)-4, h: py(25)-2,
-    menuA: ma,
-    menuB: mb,
-    draw : (itm, info, options) => {
-      g.reset().clearRect(options.x-1, options.y, options.x+options.w+1, options.y+options.h);
-      if (options.focus) g.drawRect(options.x-1, options.y, options.x+options.w+1, options.y+options.h);
-      g.setFont("Vector", py(10)).setFontAlign(-1, 0).drawString(info.text, options.x, options.y+options.h/2);
-    }
-  });
-}
-
-function saveClockinfos() {
-  data.rows = [];
-  clockinfosSettings.forEach(function(row, r) {
-    data.rows[r] = {
-      menuA: row.menuA,
-      menuB: row.menuB
-    }
-    row.remove();
-  });
-  require("Storage").writeJSON("mtnclock.json", data);
-}
-
 var drawTimeout;
 var redrawTimeout;
 
@@ -476,41 +419,6 @@ function queueDraw() {
     queueDraw();
   }, 60000 - (Date.now() % 60000));
 }
-
-function checkTouchBack(xy) {
-  return xy.x <= px(33) && xy.y < py(25);
-}
-
-function checkTouchMinus(xy) {
-  return xy.x > px(33) && xy.x < px(67) && xy.y < px(25);
-}
-
-function checkTouchPlus(xy) {
-  return xy.x >= px(67) && xy.y < px(25);
-}
-
-Bangle.on('touch', function(b, xy) {
-  // Bangle.js 2 supports long touch (type 2)
-  // On other devices, any touch will show the settings
-  if (!showingSettings && (xy.type == 2 || process.env.HWVERSION != 2)) {
-    drawClkinfoSettings();
-    showingSettings = true;
-  } else if (checkTouchBack(xy)) {
-    showingSettings = false;
-    saveClockinfos();
-    queueDraw();
-    // call setWeather after a timeout because for some reason a clockinfo
-    //  can still draw for a little bit despite calling remove() on it
-    setTimeout(setWeather, 100);
-  } else if (checkTouchMinus(xy) && clockinfosSettings.length > 0) {
-    let cl = clockinfosSettings[clockinfosSettings.length - 1];
-    cl.remove();
-    g.reset().clearRect(cl.x, cl.y, cl.x+cl.w-2, cl.y+cl.h-1);
-    clockinfosSettings.pop();
-  } else if (checkTouchPlus(xy) && clockinfosSettings.length < 3) {
-    addClockinfo(clockinfosSettings.length)
-  }
-});
 
 queueDraw();
 readWeather();
